+++ Betauer.Core/src/Deck/PlayHistory.cs
@@ -13,23 +13,23 @@ public class PlayHistory {
         Discard
     }
 
-    public class PlayedAction(int id, PlayedActionType type, IReadOnlyList<Card> cards, PokerHand? playedHand, int handScore, int gameGameScore, int totalScore ) {
+    public class PlayedAction(int id, PlayedActionType type, IReadOnlyList<Card> cards, PokerHand? playedHand, long handScore, long gameGameScore, long totalScore) {
         public int Id { get; } = id;
         public PlayedActionType Type { get; } = type;
         public List<Card> Cards { get; } = cards.Select(card => card.Clone()).ToList();
         public PokerHand? PlayedHand { get; } = playedHand;
-        public int HandScore { get; } = handScore;
-        public int GameScore { get; } = gameGameScore;
-        public int TotalScore { get; } = totalScore;
+        public long HandScore { get; } = handScore;
+        public long GameScore { get; } = gameGameScore;
+        public long TotalScore { get; } = totalScore;
     }
 
 
-    public void AddPlayAction(PokerHand hand, IReadOnlyList<Card> cards, int handScore, int score, int totalScore) {
+    public void AddPlayAction(PokerHand hand, IReadOnlyList<Card> cards, long handScore, long score, long totalScore) {
         var count = _actions.Count(ga => ga.Type == PlayedActionType.Play);
         _actions.Add(new PlayedAction(count, PlayedActionType.Play, cards, hand, handScore, score, totalScore));
     }
 
-    public void AddDiscardAction(IReadOnlyList<Card> cards, int score, int totalScore) {
+    public void AddDiscardAction(IReadOnlyList<Card> cards, long score, long totalScore) {
         var count = _actions.Count(ga => ga.Type == PlayedActionType.Discard);
         _actions.Add(new PlayedAction(count, PlayedActionType.Discard, cards, null, 0, score, totalScore));
     }
